+++ plugin-examples/src/plugins/bands-indicator/bands-indicator.ts
