+++ plugin-examples/src/plugins/bands-indicator/example/example.ts
@@ -1,6 +1,6 @@
 import { createChart } from 'lightweight-charts';
 import { generateLineData } from '../../../sample-data';
-import { BandsIndicator } from '../band-indicator';
+import { BandsIndicator } from '../bands-indicator';
 
 const chart = ((window as unknown as any).chart = createChart('chart', {
 	autoSize: true,
